# Conflicts:
#   backend-db/modules/db/couchdb/index.js

+++ web-app/src/reducers/areas/operations.js
@@ -2,18 +2,18 @@ import ApiService from '../../services/Api';
 import actions from './actions';
 import selectors from './selectors';
 import { selectors as userSels } from '../../reducers/user';
+import {
+  actions as errActions,
+} from '../../reducers/error';
+import { API_ENDPOINTS } from '../../shared/constants';
 
 const getUserAreas = ({ userId }) => async (dispatch, getState) => {
   const authUser = userSels.getProfile(getState());
   const isOwnUser = authUser && authUser.id === userId;
   dispatch(actions.getUserAreas({ userId }));
-  let response = await ApiService.get(
-    isOwnUser ? '/areas/user' : `/areas/user/${userId}`,
-  );
+  let response = await ApiService.get(`/areas/user/${userId}`);
   if (!response) {
-    response = await ApiService.get(
-      isOwnUser ? '/areas/user' : `/areas/user/${userId}`,
-    );
+    response = await ApiService.get(`/areas/user/${userId}`);
     dispatch(actions.getUserAreasError({ userId, error: true }));
     return;
   }
@@ -22,18 +22,17 @@ const getUserAreas = ({ userId }) => async (dispatch, getState) => {
   return areas;
 };
 const getAreas = () => async dispatch => {
-  dispatch(actions.getAreas());
-  let response = await ApiService.get('/areas');
-  if (!response) {
-    response = await ApiService.get('/areas');
+  try {
+    dispatch(actions.getAreas());
+    const response = await ApiService.get('/areas');
     if (!response) {
       dispatch(actions.getAreasError({ error: true }));
+      dispatch(errActions.setErrorMessage('Failed to load areas'));
       return;
     }
-  }
-  const areas = response.data;
-  try {
-    // get the other area
+    const areas = response.data;
+      // get the other area
+    /*
     const otherTraspointsResponse = await ApiService.get(
       'areas/-/trashpoints?pageSize=10&pageNumber=1',
     );
@@ -45,10 +44,11 @@ const getAreas = () => async dispatch => {
         // TODO add other count
       });
     }
+    */
+    dispatch(actions.getAreasSuccess({ areas }));
   } catch (ex) {
     console.log(ex);
   }
-  dispatch(actions.getAreasSuccess({ areas }));
 };
 
 const assignAreaLeader = ({ areaId, userId }) => async (dispatch, getState) => {
@@ -64,13 +64,31 @@ const assignAreaLeader = ({ areaId, userId }) => async (dispatch, getState) => {
     const areas = selectors.getAreas(getState());
     const area = (areas || []).find(a => a.id === areaId);
     dispatch(actions.assignAreaLeader({ areaId, area, userId }));
+    const userRes = await ApiService.get(API_ENDPOINTS.FETCH_USER_BY_ID(userId));
+    if (!userRes) {
+      dispatch({ type: 'GET_USER_ERROR' });
+    } else {
+      dispatch({
+        type: 'GET_USER_SUCCESS',
+        payload: { user: userRes.data },
+      });
+    }
   }
 };
 
 const removeAreaLeader = ({ areaId, userId }) => async dispatch => {
-  const response = await ApiService.delete(`/areas/${areaId}/leader`);
+  const response = await ApiService.delete(`/areas/${areaId}/leader/${userId}`);
   if (response) {
     dispatch(actions.removeAreaLeader({ areaId, userId }));
+    const userRes = await ApiService.get(API_ENDPOINTS.FETCH_USER_BY_ID(userId));
+    if (!userRes) {
+      dispatch({ type: 'GET_USER_ERROR' });
+    } else {
+      dispatch({
+        type: 'GET_USER_SUCCESS',
+        payload: { user: userRes.data },
+      });
+    }
   }
 };
 export default {
